fix: Small tweaks
